Move out of apibindings to bindings, and apply vanity URL
