+++ bindings/go/internal/testhelper/errcontainsstr.go
