+++ bindings/go/osvdev/models.go
