+++ bindings/go/osvdev/osvdev_test.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
-	"github.com/google/osv.dev/apibindings/go/internal/testhelper"
-	"github.com/google/osv.dev/apibindings/go/osvdev"
 	"github.com/ossf/osv-schema/bindings/go/osvschema"
+	"osv.dev/bindings/go/internal/testhelper"
+	"osv.dev/bindings/go/osvdev"
 )
 
 func TestOSVClient_GetVulnsByID(t *testing.T) {
